requirements.txt
# NDA Analyzer - Python Dependencies
# Updated: January 2025

# Web framework
streamlit>=1.32.0

# Environment management
python-dotenv>=1.0.0

# Document processing
PyPDF2>=3.0.0
python-docx>=1.0.0

# ML/Embeddings
sentence-transformers>=2.5.0
faiss-cpu>=1.7.4
numpy>=1.24.0

# PyTorch (CPU-only for smaller image size)
torch>=2.2.0
--extra-index-url https://download.pytorch.org/whl/cpu

# HTTP client
requests>=2.31.0

# Data processing
pandas>=2.2.0

# Optional: OpenAI SDK for cleaner API calls
# openai>=1.0.0
